fix build

+++ src/main/java/com/blubb/lavf/AVStream.java
@@ -3,7 +3,7 @@ package com.blubb.lavf;
 public class AVStream {
 	long avstream;
 
-	public AVStream(int avstream) {
+	public AVStream(long avstream) {
 		this.avstream = avstream;
 	}
 }
