+++ src/main/java/com/blubb/lavf/LAVFNative.java
@@ -77,7 +77,7 @@ public class LAVFNative {
 
 	native long streamts_to_basets(long fmt_ctx, long ts);
 
-	native int av_add_video_stream(long fmt_ctx, int codecId, int width, int height, int bitrate, int frametime, int pixfmt);
+	native long av_add_video_stream(long fmt_ctx, int codecId, int width, int height, int bitrate, int frametime, int pixfmt);
 
     native int av_open_video(long fmt_ctx, long stream);
 }
